Rename org.wildfly.swarm.microprofile.metrics package in the test suite. Fixes #89
