+++ testsuite/extra/src/main/java/io/smallrye/metrics/test/MetricsSummary.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.microprofile.metrics;
+package io.smallrye.metrics.test;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
