+++ testsuite/extra/src/test/java/io/smallrye/metrics/test/inject/NonReusableMetricInjectionBean.java
@@ -15,7 +15,7 @@
  *   limitations under the License.
  */
 
-package org.wildfly.swarm.microprofile.metrics.inject;
+package io.smallrye.metrics.test.inject;
 
 import org.eclipse.microprofile.metrics.Counter;
 import org.eclipse.microprofile.metrics.MetricUnits;
