+++ testsuite/extra/src/test/java/io/smallrye/metrics/test/reusability/ReusableMetricWithDifferingTagsBean.java
@@ -15,23 +15,26 @@
  *   limitations under the License.
  */
 
-package org.wildfly.swarm.microprofile.metrics.reusability;
+package io.smallrye.metrics.test.reusability;
 
 import org.eclipse.microprofile.metrics.annotation.Counted;
 
-import javax.enterprise.context.ApplicationScoped;
+public class ReusableMetricWithDifferingTagsBean {
 
-@ApplicationScoped
-public class NonReusableMetricBean {
-
-    @Counted(name = "countedMethod", absolute = true)
-    public void countedMethodOne() {
-
+    @Counted(name = "colorCounter", absolute = true, tags = {"color=blue"}, reusable = true)
+    public void colorBlue1() {
     }
 
-    @Counted(name = "countedMethod", absolute = true)
-    public void countedMethodOne2() {
+    @Counted(name = "colorCounter", absolute = true, tags = {"color=red"}, reusable = true)
+    public void colorRed1() {
+    }
 
+    @Counted(name = "colorCounter", absolute = true, tags = {"color=blue"}, reusable = true)
+    public void colorBlue2() {
+    }
+
+    @Counted(name = "colorCounter", absolute = true, tags = {"color=red"}, reusable = true)
+    public void colorRed2() {
     }
 
 }
